pyproject.toml
[project]
name = "signaltower"
version = "0.1.0"
description = "FastAPI service that controls a Velleman K8055 signal tower via USB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyusb>=1.2.1",
]

[project.scripts]
signaltower = "signaltower.app:main"

[tool.hatch.build.targets.wheel]
packages = ["signaltower"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
